+++ js/data.js
@@ -1,0 +1,29 @@
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Иван',
+  'Мария',
+  'Артем',
+  'Ольга',
+  'Петр',
+  'Наталья',
+];
+
+const DESCRIPTIONS_TEXTS = [
+  'Фото кота в лунном свете.',
+  'Кот среди подсолнухов. Подражание ван Гогу.',
+  'Экзистенциальные основы бытия кошачьей жизни.',
+  'Кот моей бабушки.',
+  'Я и мой кот.',
+  'Кот в поисках смысла.',
+  'Моя любимая кошка.',
+];
+
+export {MESSAGES,NAMES,DESCRIPTIONS_TEXTS};
