+++ js/utils.js
@@ -1,0 +1,17 @@
+/* source - https://learn.javascript.ru/task/random-int-min-max */
+const getRandomInteger = (min, max) => {
+
+  min = Math.abs(min);
+  max = Math.abs(max);
+
+  if (max <= min) {
+    return min;
+  }
+  return Math.floor(min + Math.random() * (max + 1 - min));
+};
+
+const checkStringLength = (stringToCheck, maxLength) => stringToCheck.length <= maxLength;
+
+const generateRandomId = () => getRandomInteger(0,1000) + new Date().getTime();
+
+export {getRandomInteger,checkStringLength,generateRandomId};
